Add files via upload
This program is linked to the program "pizza slices.c". It asks the user the diameter of the pizza they ordered as well as the number of people attending their party, and then calculates the number of pizzas to be purchased for them based on the assumption that each person will have four slices. And the slices have the same area.